pyproject.toml
[project]
name = "my-game"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
  "altgraph==0.17.4",
  "glfw>=2.10.0",
  "numpy==2.2.3",
  "packaging==24.2",
  "pefile==2023.2.7",
  "pillow>=11.3.0",
  "pygame==2.6.1",
  "pygame-ce==2.5.3",
  "pygame-gui==0.6.8",
  "pyinstaller==6.12.0",
  "pyinstaller-hooks-contrib==2025.1",
  "pyopengl>=3.1.10",
  "pyopengl-accelerate>=3.1.10",
  "python-i18n==0.3.9",
  "pywin32-ctypes==0.2.3",
  "setuptools==80.8.0",
  "six==1.17.0",
]
